moved functions from avatarHeader to Home
