+++ resources/js/components/pages/Home.js
@@ -1,12 +1,90 @@
 import React from "react";
+import { useRef, useState } from "react";
 import AvatarHeader from "../layouts/AvatarHeader";
 import Header from "../layouts/Header";
 
 const Home = () => {
+
+    var hpTotal = 50;
+    var xpTotal = 50;
+    const [hp, setHp] = useState(50);
+    const [xp, setXp] = useState(0);
+    const [hpBarWidth, sethpBarWidth] = useState("100");
+    const [hpHitWidth, sethpHitWidth] = useState("0");
+    const [HpIncreaseWidth, setHPIncreaseWidth] = useState("0");
+
+    const [xpBarWidth, setXPBarWidth] = useState("0");
+    const [xpIncreaseWidth, setXPIncreaseWidth] = useState("0");
+
+    const hitHandler = () => {
+        let updatedHp;
+
+        if(hp == 0){
+            updatedHp = 50;
+        }else{
+            updatedHp = hp - 10;
+        }
+        setHp(updatedHp);
+
+        const newhpBarWidth = updatedHp / hpTotal * 100;
+        const newhpHitWidth = 10 / hp * 100;
+
+        sethpHitWidth(newhpHitWidth);
+
+        setTimeout(function(){
+            sethpHitWidth(0);
+            sethpBarWidth(newhpBarWidth);
+        }, 500);
+    }
+
+    const healHandler = () => {
+        let updatedHp;
+        let newHPBarWidth;
+
+        if(hp == hpTotal){
+            //    full health modal or sumn
+            updatedHp = hpTotal;
+        }else{
+            updatedHp = hp + 10;
+            newHPBarWidth = updatedHp / hpTotal * 100;
+        }
+        setHp(updatedHp);
+
+        setHPIncreaseWidth(newHPBarWidth);
+
+        setTimeout(function(){
+            sethpBarWidth(newHPBarWidth);
+        }, 500);
+    }
+
+    const addXPHandler = () => {
+        let updatedXp = 0;
+        let newXPBarWidth;
+        if(xp == xpTotal){
+            newXPBarWidth = 0;
+        }else{
+            updatedXp = xp + 10;
+            newXPBarWidth = updatedXp / xpTotal * 100;
+        }
+        setXp(updatedXp);
+
+        setXPIncreaseWidth(newXPBarWidth);
+
+        setTimeout(function(){
+            setXPBarWidth(newXPBarWidth);
+        }, 500);
+
+    }
     return (
         <div className="internal-pages">
             <Header page="home"/>
-            <AvatarHeader/>
+            <AvatarHeader hp={hp} hpTotal={hpTotal} hpBarWidth={hpBarWidth} hpHitWidth={hpHitWidth} HpIncreaseWidth={HpIncreaseWidth} xp={xp} xpTotal={xpTotal} xpBarWidth={xpBarWidth} xpIncreaseWidth={xpIncreaseWidth}/>
+             {/* TESTING */}
+             <div style={{margin: 15 + "%"}}>
+                <button style={{margin: 10 + "px"}} className="btn btn-danger" onClick={hitHandler}>damage</button>
+                <button  style={{margin: 10 + "px"}} className="btn btn-success" onClick={healHandler}>heal</button>
+                <button  style={{margin: 10 + "px"}} className="btn btn-primary" onClick={addXPHandler}>add XP</button>
+             </div>
         </div>
     );
 }
